Cookie implementation at customerList, staffList

+++ LouisJ/build/web/staffList.jsp
@@ -7,6 +7,8 @@
 <%@page import="java.util.List"%>
 <%@page import="model.Employee"%>
 <%@page contentType="text/html" pageEncoding="UTF-8"%>
+<jsp:useBean id="sessionUtil" class="util.UserSessionUtil" />
+<jsp:useBean id="user" class="interfaces.User" />
 <!DOCTYPE html>
 <html>
     <head>
@@ -19,7 +21,9 @@
         <script src="https://kit.fontawesome.com/a293bfc92d.js" crossorigin="anonymous"></script>
     </head>
     <body>
-        <a class="btn btn-primary" href="" style="width: 120px; margin: 2rem 2rem 0 2rem">Add Staff</a>
+        <% if (sessionUtil.getCurrentLoginUser(request.getCookies()).equals("manager")){ %>
+            <a class="btn btn-primary" href="" style="width: 120px; margin: 2rem 2rem 0 2rem">Add Staff</a>
+        <% } %>
         
         <div class="row">
             <% 
@@ -46,15 +50,15 @@
                         </div>
                     </div>
                     <div class="card-footer p-3">
-                        <a href="" style="text-decoration: none;">
-                            <i class="fa-regular fa-pen-to-square fa-2xl"></i>
-                        </a>
-                        <a href="" class="btn btn-sm btn-primary">
-                            <i class="fas fa-user fa-sm"></i>
-                        </a>
-                        <a id="deleteItem" style="cursor: pointer;" class="deleteItem" data-target="#basic" data-toggle="modal">
-                            <i class="fa fa-trash fa-2xl"></i>
-                        </a>
+                        
+                        <% if (sessionUtil.getCurrentLoginUser(request.getCookies()).equals("manager")){ %>
+                            <a href="" style="text-decoration: none;">
+                                <i class="fa-regular fa-pen-to-square fa-2xl"></i>
+                            </a>
+                            <a id="deleteItem" style="cursor: pointer;" class="deleteItem" data-target="#basic" data-toggle="modal">
+                                <i class="fa fa-trash fa-2xl"></i>
+                            </a>
+                        <% } %>
                     </div>
                 </div>
             </div>
